Updated Wordbank to use WordCategory class
